Scroll to active sidebar item on page load
When loading a page that is farther down the sidebar (like a tutorial), that area of the sidebar isn't visible on page load. This makes it hard to know what sections are in the tutorial.

Now, if a sidebar section is outside the visible window but is active on page load the sidebar is scrolled to this section.